修改飞船设置返回值

+++ capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/pojo/PkCopyBase.java
@@ -69,7 +69,7 @@ public class PkCopyBase extends Model<PkCopyBase> {
     /**
      * 命中率
      */
-    private Integer hitRate;
+    private Double hitRate;
     /**
      * 机动力
      */
